backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz #timezome
sqlparse

psycopg[binary]
python-dotenv
python-decouple
